Fixes #1 Clean test structure, add rakefile, enable xml reports
