+++ examples/quickapp/tests/fute/.gitignore
@@ -1,1 +1,2 @@
 *.log
+results
